styles chat message box
