+++ client/src/pages/SignUp.jsx
@@ -44,7 +44,7 @@ function SignUp() {
 			});
 			console.log(response);
 			if (response.ok) {
-				window.location.href = "/websockettest";
+				window.location.href = "/globalchat";
 			} else {
 				const errorData = await response.json();
 				console.log(errorData.error.errors);
